Fix a bunch of compiler warnings, like C4127 resulting from do {} while(0) in MSVC.

+++ src/Math/MathFunc.h
@@ -426,7 +426,7 @@ bool EqualUlps(float a, float b, int maxUlps = 10000);
 template<typename T> inline bool IsFinite(const T &value) { return true; }
 template<> inline bool IsFinite<float>(const float &value) { return isfinite(value) != 0; }
 template<> inline bool IsFinite<double>(const double &value) { return isfinite(value) != 0; }
-template<> inline bool IsFinite<long double>(const long double &value) { return isfinite(value) != 0; }
+template<> inline bool IsFinite<long double>(const long double &value) { return isfinite((double)value) != 0; }
 
 /// Returns true if the given value is +inf or -inf.
 float IsInf(float value);
